style: project setting member page revamp
